+++ app/src/androidTest/java/edu/monash/student/happyactive/data/ActivitySessionDaoTest.java
@@ -1,0 +1,4 @@
+package edu.monash.student.happyactive.data;
+
+public class ActivitySessionDaoTest {
+}
