+++ app/src/main/java/edu/monash/student/happyactive/ActivityPackages/Repositories/ActivitySessionRepository.java
@@ -1,0 +1,4 @@
+package edu.monash.student.happyactive.ActivityPackages.Repositories;
+
+public class ActivitySessionRepository {
+}
