+++ app/src/main/java/edu/monash/student/happyactive/ActivityPackages/viewModels/ActivitySessionViewModel.java
@@ -1,4 +1,4 @@
-package edu.monash.student.happyactive.ActivityPackages;
+package edu.monash.student.happyactive.ActivityPackages.viewModels;
 
 import android.app.Application;
 
@@ -9,15 +9,17 @@ import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 
 import java.util.Date;
-import java.util.List;
 
+import edu.monash.student.happyactive.ActivityPackages.ActivityPackageRepository;
+import edu.monash.student.happyactive.ActivityPackages.ActivityPackageStatus;
+import edu.monash.student.happyactive.ActivityPackages.PackageSessionManager;
 import edu.monash.student.happyactive.data.entities.ActivityPackage;
 import edu.monash.student.happyactive.data.entities.ActivitySession;
 import edu.monash.student.happyactive.data.entities.Task;
 import edu.monash.student.happyactive.data.relationships.ActivityPackageWithTasks;
 
 public class ActivitySessionViewModel  extends AndroidViewModel {
-    private  ActivityPackageRepository activityPackageRepository;
+    private ActivityPackageRepository activityPackageRepository;
     private LiveData<ActivityPackageWithTasks> acitivtyPackageWithTasks;
     private ActivitySession activitySession;
     private PackageSessionManager sessionManager;
